Renamed the files according to the style guidelines
