+++ src/library/borrowbook/BorrowBookUI.java
@@ -13,12 +13,12 @@ public class BorrowBookUI {
 	
 	public static enum uI_STaTe { INITIALISED, READY, RESTRICTED, SCANNING, IDENTIFIED, FINALISING, COMPLETED, CANCELLED };
 
-	private bORROW_bOOK_cONTROL CoNtRoL;
+	private BorrowBookControl CoNtRoL;
 	private Scanner InPuT;
 	private uI_STaTe StaTe;
 
 	
-	public BorrowBookUI(bORROW_bOOK_cONTROL control) {
+	public BorrowBookUI(BorrowBookControl control) {
 		this.CoNtRoL = control;
 		InPuT = new Scanner(System.in);
 		StaTe = uI_STaTe.INITIALISED;
